Port to Linux (GCC 13, Bazel 8)
1. Add bin/bazel wrapper for bazelisk with writable cache directories
2. Fix missing includes: <cstring>, <cmath>, <cstdint>, <functional>,
   <cstddef>, <unistd.h>
3. Use compiler-conditional warning flags via select() for Clang vs GCC
4. Fix memset/memcpy on non-trivial types with static_cast<void*>
5. Fix _Static_assert portability with PENTAGO_STATIC_ASSERT macro
6. Use tfm::format instead of bare format
7. Fix operator precedence, dangling-else braces, sign-compare
8. Update MPI detection for Linux (.so, /opt/amazon/openmpi/)
9. Upgrade googletest to v1.15.2
10. Switch lzma URL from tukaani.org to GitHub
11. Add zlib-1.3.1 hash to supertensor_test
12. Add Claude Code config, CLAUDE.md, .mcp.json, .bazelignore

+++ pentago/mpi/requests.cc
@@ -107,7 +107,7 @@ void requests_t::cancel_and_waitall() {
   GEODE_ASSERT(!immediate_count);
 #endif
   callbacks.clear();
-  for (int i=0;i<requests.size();i++) {
+  for (size_t i=0;i<requests.size();i++) {
     GEODE_ASSERT(cancellables[i]);
     CHECK(MPI_Cancel(&requests[i]));
   }
